Extract features for all valid music
